[CFE] Make strong tests verification skip platform
Currently - i.e. without this CL - (almost) every test run by the strong
test suite verifies the platform libraries as well. That's basically a
waste of time.
This CL makes the suite skip it for the platform.

Runtimes changes (on my computer) as follows:

With -DskipVm=true:
With this CL: ~1:55
Without/before this CL: ~8:30

Without -DskipVm=true:
With this CL: ~4:35
Without/before this CL: ~11:30

Change-Id: I6c46b86935bea29911775cbc0544ea7a0b60b779
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/180562
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>