[analyzer/ffi] Fix `null` exceptions
The analyzer has not been migrated, so this is based on manual
inspecation of the API about what can be `null` for now.

Note with the analyzer from master (running in nnbd by default) this
file contains many nullable to non-null assignments. However, we cannot
migrate this file to nnbd yet because of the language version it's used
in.

Closes: https://github.com/dart-lang/sdk/issues/44763

Change-Id: If0432341e7fefb973b0229bcc325eeefed84f7cb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/180821
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>
Commit-Queue: Daco Harkes <dacoharkes@google.com>