Analysis server: Remove unnecessary imports; standardize on prefix use
In a few files, an element is available twice: once without a prefix,
and once with a prefix, and the element is referenced in both ways. This
triggers an "unnecessary import" diagnostic, and changing all of the
references to be unified resolves the issue.

Bug: https://github.com/dart-lang/sdk/issues/44569
Change-Id: I3c5a3f397f0a3603624a2d23becfda2753cc9f8a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/177280
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>