[dart2js] Ensure that the class member environment for [node] is computed for live members
This CL ensures that the member environment for the enclosing class
of a live member is computed.

Static members and mixed in members a member can be become live
through static access and mixin application, respectively, which does
not require lookup into the class members.

Since the J-model class environment is computed from the K-model
environment, not ensuring the computation of the class members, can
result in a live member being present in the J-model but unavailable
when queried as a member of its enclosing class.

Closes #44754

Change-Id: I5a735f4605737947cca082450803c35887b0b7a0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/181801
Commit-Queue: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Sigmund Cherem <sigmund@google.com>