[gardening] Small partial revert that enabled validating constants after reload
While addresing a comment in the review of [0], the CL removed a bailout in
`Isolate::ValidateConstants()` that disabled validating of constants
when hot-reload was performed. Apparently there is still an issue with
that. So this CL adds this bailout again.

[0] https://dart-review.googlesource.com/c/sdk/+/182561

TEST=Partial revert, fixes constant validation failures.

Issue https://github.com/dart-lang/sdk/issues/44862

Change-Id: Iefe9acceb344e73fad97246064e91a32a00dad10
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/183002
Commit-Queue: Martin Kustermann <kustermann@google.com>
Reviewed-by: Clement Skau <cskau@google.com>