Add the fix data file to analyzableFilePatterns
Devon noticed the other day that this was missing. I was curious so I
commented out all of the non-Dart file patterns and ran the tests. There
were no failures. There were failures when I also commented out the
file pattern for Dart files, so that much is being used. My guess is
that it would be better to remove the seemingly unused patterns, but
then I have to ask whether we shouldn't just remove this field and
replace uses by a check that the name ends in '.dart'. Let me know what
you think.

Change-Id: Icda44a6fa532989f86cdb52acb3bd0f7a5b28e20
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/184900
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>